+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Show why a modules is installed in module info.
 - On app-graph, create SVG next to PNG and graph file.
 - Add command app-module-config-uncache.
+- Add command app-module-config-validate.
 
 ## v1.0.2
 - Fix structure migration bug in module update.